requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.3.1
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
mock==4.0.3
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
Pyphen==0.10.0
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
textstat==0.7.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
